requirements.txt
MouseInfo==0.1.3
mss==6.0.0
numpy==1.19.1
opencv-python==4.4.0.40
Pillow==7.2.0
pkg-resources==0.0.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python3-xlib==0.15
PyTweening==1.0.3
